Introduced TestCase and ITestCase,
which will replace part of the role currently played by FirstClassCommand.
The reason for this is that it turns out that the xUnit.net framework
makes certain assumptions about the IMethodInfo instance passed to the
TestCommand constructor. At the time a static test method starts
enumerating test cases, there's not enough information available about the
actual test method.

Instead, the new approach will be to have test methods yield sequences of
ITestCase, and the ITestCase interface will then subsequently have a
conversion method that returns a proper ITestCommand.